requirements.txt
pyodbc==4.0.30
azure-identity
azure-keyvault-secrets
